added new starter screen and changed example starter code to make it simpler.
