+++ app/src/main/java/com/example/planetZe_gp5/HomeFragment.java
@@ -5,6 +5,8 @@ import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 import android.widget.Button;
+import android.widget.TextView;
+
 import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
 import androidx.fragment.app.Fragment;
@@ -17,8 +19,6 @@ public class HomeFragment extends Fragment {
         View view = inflater.inflate(R.layout.activity_home_fragment, container, false);
 
         Button buttonRecyclerView = view.findViewById(R.id.buttonRecyclerView);
-        Button buttonScroller = view.findViewById(R.id.buttonScroller);
-        Button buttonSpinner = view.findViewById(R.id.buttonSpinner);
         Button buttonManageItems = view.findViewById(R.id.buttonManageItems);
 
         buttonRecyclerView.setOnClickListener(new View.OnClickListener() {
@@ -28,24 +28,18 @@ public class HomeFragment extends Fragment {
             }
         });
 
-        buttonScroller.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View v) {
-                loadFragment(new ScrollerFragment());
-            }
-        });
-
-        buttonSpinner.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View v) {
-                loadFragment(new SpinnerFragment());
-            }
-        });
-
         buttonManageItems.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View v) { loadFragment(new ManageItemsFragment());}
         });
+
+        TextView textView = view.findViewById(R.id.textView);
+        StringBuilder text = new StringBuilder();
+        text.append("Scroller example\n");
+        for (int i = 0; i < 100; i++) {
+            text.append("Line ").append(i + 1).append("\n");
+        }
+        textView.setText(text.toString());
 
         return view;
     }
