+++ app/src/main/java/com/example/planetZe_gp5/MainActivity.java
@@ -19,7 +19,7 @@ public class MainActivity extends AppCompatActivity {
 //        dbModel.writeData("testDemo/movies", "B07 Demo!");
 
         if (savedInstanceState == null) {
-            loadFragment(new HomeFragment());
+            loadFragment(new StartFragment());
         }
     }
